requirements.txt
langchain
langchain-openai
langchain-community
langchainhub
python-dotenv